+++ src/main/java/com/knu/sosuso/capstone/dto/oauth2/CustomOAuth2User.java
@@ -1,6 +1,5 @@
-package com.knu.sosuso.capstone.dto;
+package com.knu.sosuso.capstone.dto.oauth2;
 
-import com.knu.sosuso.capstone.dto.request.GoogleUserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
