+++ src/main/java/com/knu/sosuso/capstone/dto/oauth2/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.knu.sosuso.capstone.dto.request;
+package com.knu.sosuso.capstone.dto.oauth2;
 
 public record GoogleUserInfo(
         String sub,
