+++ src/main/java/com/knu/sosuso/capstone/service/CustomOAuth2UserService.java
@@ -1,8 +1,8 @@
 package com.knu.sosuso.capstone.service;
 
 import com.knu.sosuso.capstone.domain.User;
-import com.knu.sosuso.capstone.dto.CustomOAuth2User;
-import com.knu.sosuso.capstone.dto.request.GoogleUserInfo;
+import com.knu.sosuso.capstone.dto.oauth2.CustomOAuth2User;
+import com.knu.sosuso.capstone.dto.oauth2.GoogleUserInfo;
 import com.knu.sosuso.capstone.dto.response.GoogleResponse;
 import com.knu.sosuso.capstone.dto.response.OAuth2Response;
 import com.knu.sosuso.capstone.repository.UserRepository;
